govt_data_pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
langchain>=0.0.300
langchain-pinecone>=0.0.1
langchain-openai>=0.0.2
langchain-anthropic>=0.0.6
pinecone-client>=3.0.0
openai>=1.0.0
anthropic>=0.8.0